Publish testing module internally
