+++ testing/shared/src/main/scala/org/http4s/Http4sSuite.scala
@@ -26,7 +26,7 @@ import org.scalacheck.Prop
 
 /** Common stack for http4s' munit based tests
   */
-trait Http4sSuite
+private[http4s] trait Http4sSuite
     extends CatsEffectSuite
     with DisciplineSuite
     with munit.ScalaCheckEffectSuite
